+++ all-is-cubes-ui/src/vui/widget_trait.rs
@@ -310,8 +310,7 @@ pub(crate) fn instantiate_widget<W: Widget + 'static>(
     use all_is_cubes::transaction::Transaction as _;
 
     let mut space = Space::builder(grant.bounds).build();
-    let txn = vui::install_widgets(grant, &vui::LayoutTree::leaf(Arc::new(widget)))
-        .expect("widget instantiation");
+    let txn = vui::install_widgets(grant, &vui::leaf_widget(widget)).expect("widget instantiation");
     txn.execute(&mut space, &mut transaction::no_outputs)
         .expect("widget transaction");
     (txn.bounds_only_cubes(), space)